Implement creation of JWTs for distributed auth
The endpoint /auth/token will create a JWT signed by the viewer with
the currently logged-in user detailed. The user can then send this
token to another CUDL service independent of the viewer, which can
verify that the token's signature is valid (and has not expired), and
as a result know that the requester can act as the claimed user.

The intention is to use this initially when breaking out the tagging
functionality from the viewer.

To get a token, POST to
    /auth2/token?audience=http://services.cudl.lib.cam.ac.uk
The audience URL should identify the service that will receive the
token. The token will expire after 30 seconds, so will need to be
regularly refreshed by the logged-in user.