Configure docker tomcat memory usage
Current JVM versions are able to automatically detect memory limits of
a container they're running in. Default is currently to use 25% of max
memory.

This limits Tomcat's container to 1G and has it use at most 85% of
that. This seems to slightly reduce startup time (which is currently
really slow - about 1 minute).